[Bugfix] New implementation of naw_securedl hook
Hook is now FALL aware and checks sys_file.fe_groups field
added by Media extension.
There is a bug in naw_securedl that makes it impossible
to add a ?1234567890 to the image url. So when naw_securedl is installed
we don't add the timestamp in FE when using the thumnailViewhelper.

Fixes: #48044
Change-Id: I7f5ca4407f7f3eaf654442a0cf0a0ed489087c14