[BUGFIX] Decouple metadata output from thumbnail service
Currently metadata are displayed by the thumbnail service.
This happen when the thumbnail is requested with option "wrap".
If that is OK for a first implementation in the BE module, this is
not desired as default output.

The change also remove some old code which was there for working
round a bug. Good for us, the bug is resolved:
https://review.typo3.org/#/c/20439

Change-Id: Ibca1db0a071d392fba94ba5b9396c11affbee1ba
Fixes: #49683