[BUGFIX] Introduce own Be/ContainerViewHelper
Currently the ContainerViewHelper of ext:extensionmanager is
used. But in TYPO3 6.2 the viewHelper is removed.

Add a own ContainerViewHelper to ext:media to be compatible
with 6.1 and 6.2.
Added a copy of the new ext:fluid Be/ContainerViewHelper so when
we raise version dependicies we uses can remove this one in
faivor of the one from core.

Fixes: #50602
Releases: 1.1.0
Change-Id: I09386cd56c8775b36577871467d060e079d67a70