[BUGFIX] FilePicker lacks support for file field
The FilePicker didn't support file fields but rather only inline fields.

Now it is detecting if the flex form field is a inline relation or file field and based on this another javascript function defined in the core´s formEngine is used to fill in form values.

Since the standard file field uses the files publicUrl as value we had to introduce a data-url attribute on imageViewHelper. Having this attribute on one of the links for a record is enough.